// src/me-client-ts/dymensionxyz.dymension.eibc/types/dymensionxyz/dymension/eibc/demand_order.ts
export interface DemandOrder {
  /**
   * id is a hash of the form generated by GetRollappPacketKey,
   * e.g status/rollappid/packetProofHeight/packetDestinationChannel-PacketSequence which guarantees uniqueness
   */
  id: string
  /**
   * tracking_packet_key is the key of the packet that is being tracked.
   * This key can change depends on the packet status.
   */
  trackingPacketKey: string
  /** price is the amount that the fulfiller sends to original eibc transfer recipient */
  price: Coin[]
  /** fee is the effective profit made by the fulfiller because they pay price and receive fee + price */
  fee: Coin[]
  recipient: string
  trackingPacketStatus: Status
  rollappId: string
  type: RollappPacket_Type
  /** fulfiller_address is the bech32-encoded address of the account which fulfilled the order. */
  fulfillerAddress: string
}

function createBaseDemandOrder(): DemandOrder {
  return {
    id: '',
    trackingPacketKey: '',
    price: [],
    fee: [],
    recipient: '',
    trackingPacketStatus: 0,
    rollappId: '',
    type: 0,
    fulfillerAddress: '',
  }
}

export const DemandOrder = {
  encode(message: DemandOrder, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.id !== '') {
      writer.uint32(10).string(message.id)
    }
    if (message.trackingPacketKey !== '') {
      writer.uint32(18).string(message.trackingPacketKey)
    }
    for (const v of message.price) {
      Coin.encode(v!, writer.uint32(26).fork()).ldelim()
    }
    for (const v of message.fee) {
      Coin.encode(v!, writer.uint32(34).fork()).ldelim()
    }
    if (message.recipient !== '') {
      writer.uint32(42).string(message.recipient)
    }
    if (message.trackingPacketStatus !== 0) {
      writer.uint32(64).int32(message.trackingPacketStatus)
    }
    if (message.rollappId !== '') {
      writer.uint32(74).string(message.rollappId)
    }
    if (message.type !== 0) {
      writer.uint32(80).int32(message.type)
    }
    if (message.fulfillerAddress !== '') {
      writer.uint32(90).string(message.fulfillerAddress)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): DemandOrder {
    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseDemandOrder()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          message.id = reader.string()
          break
        case 2:
          message.trackingPacketKey = reader.string()
          break
        case 3:
          message.price.push(Coin.decode(reader, reader.uint32()))
          break
        case 4:
          message.fee.push(Coin.decode(reader, reader.uint32()))
          break
        case 5:
          message.recipient = reader.string()
          break
        case 8:
          message.trackingPacketStatus = reader.int32() as any
          break
        case 9:
          message.rollappId = reader.string()
          break
        case 10:
          message.type = reader.int32() as any
          break
        case 11:
          message.fulfillerAddress = reader.string()
          break
        default:
          reader.skipType(tag & 7)
          break
      }
    }
    return message
  },

  fromJSON(object: any): DemandOrder {
    return {
      id: isSet(object.id) ? String(object.id) : '',
      trackingPacketKey: isSet(object.trackingPacketKey) ? String(object.trackingPacketKey) : '',
      price: Array.isArray(object?.price) ? object.price.map((e: any) => Coin.fromJSON(e)) : [],
      fee: Array.isArray(object?.fee) ? object.fee.map((e: any) => Coin.fromJSON(e)) : [],
      recipient: isSet(object.recipient) ? String(object.recipient) : '',
      trackingPacketStatus: isSet(object.trackingPacketStatus)
        ? statusFromJSON(object.trackingPacketStatus)
        : 0,
      rollappId: isSet(object.rollappId) ? String(object.rollappId) : '',
      type: isSet(object.type) ? rollappPacket_TypeFromJSON(object.type) : 0,
      fulfillerAddress: isSet(object.fulfillerAddress) ? String(object.fulfillerAddress) : '',
    }
  },

  toJSON(message: DemandOrder): unknown {
    const obj: any = {}
    message.id !== undefined && (obj.id = message.id)
    message.trackingPacketKey !== undefined && (obj.trackingPacketKey = message.trackingPacketKey)
    if (message.price) {
      obj.price = message.price.map((e) => (e ? Coin.toJSON(e) : undefined))
    } else {
      obj.price = []
    }
    if (message.fee) {
      obj.fee = message.fee.map((e) => (e ? Coin.toJSON(e) : undefined))
    } else {
      obj.fee = []
    }
    message.recipient !== undefined && (obj.recipient = message.recipient)
    message.trackingPacketStatus !== undefined &&
      (obj.trackingPacketStatus = statusToJSON(message.trackingPacketStatus))
    message.rollappId !== undefined && (obj.rollappId = message.rollappId)
    message.type !== undefined && (obj.type = rollappPacket_TypeToJSON(message.type))
    message.fulfillerAddress !== undefined && (obj.fulfillerAddress = message.fulfillerAddress)
    return obj
  },

  fromPartial<I extends Exact<DeepPartial<DemandOrder>, I>>(object: I): DemandOrder {
    const message = createBaseDemandOrder()
    message.id = object.id ?? ''
    message.trackingPacketKey = object.trackingPacketKey ?? ''
    message.price = object.price?.map((e) => Coin.fromPartial(e)) || []
    message.fee = object.fee?.map((e) => Coin.fromPartial(e)) || []
    message.recipient = object.recipient ?? ''
    message.trackingPacketStatus = object.trackingPacketStatus ?? 0
    message.rollappId = object.rollappId ?? ''
    message.type = object.type ?? 0
    message.fulfillerAddress = object.fulfillerAddress ?? ''
    return message
  },
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends Array<infer U>
    ? Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>

type KeysOfUnion<T> = T extends T ? keyof T : never
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never }